+++ ai/core/src/unitycatalog/ai/core/utils/validation_utils.py
@@ -8,7 +8,7 @@ from unitycatalog.ai.core.utils.type_utils import is_time_type
 if TYPE_CHECKING:
     from databricks.sdk.service.catalog import FunctionInfo
 
-OSS_MAX_COMMENT_OR_NAME_LENGTH = 255
+OSS_MAX_FUNCTION_NAME_LENGTH = 255
 
 
 class FullFunctionName(NamedTuple):
@@ -135,8 +135,8 @@ def validate_function_name_length(function_name: str) -> None:
         function_name: The name of the function being created within Unity Catalog.
     """
     name_length = len(function_name)
-    if name_length > OSS_MAX_COMMENT_OR_NAME_LENGTH:
+    if name_length > OSS_MAX_FUNCTION_NAME_LENGTH:
         raise ValueError(
-            f"The maximum length of a function name is {OSS_MAX_COMMENT_OR_NAME_LENGTH}. "
+            f"The maximum length of a function name is {OSS_MAX_FUNCTION_NAME_LENGTH}. "
             f"The name supplied is {name_length} characters long."
         )
